requirements.txt
apertium-apy==0.12.0
appnope==0.1.4
asttokens==2.4.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
distlib==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
fonttools==4.49.0
gunicorn==21.2.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
wordcloud==1.9.3
zipp==3.17.0
